requirements.txt
## ansible runner
ansible_runner
requests
ansible-core
pyyaml